PYIC-5624: Move sniffing to form submission handler
