+++ src/app/ipv/middleware.test.js
@@ -1038,32 +1038,6 @@ describe("journey middleware", () => {
       expect(res.render).to.not.have.been.called;
       expect(next).to.have.been.calledOnce;
     });
-
-    it("should include dcmaw event when not smartphone for re-render", async function () {
-      req.method = "POST";
-      req.body.journey = undefined;
-      req.isSmartphoneUser = false;
-
-      await middleware.formRadioButtonChecked(req, res, next);
-
-      expect(res.render).to.have.been.calledWith(
-        `ipv/page/page-ipv-identity-document-start.njk`,
-        sinon.match.has("dcmawEvent", "dcmaw"),
-      );
-    });
-
-    it("should include smartphone event when smartphone for re-render", async function () {
-      req.method = "POST";
-      req.body.journey = undefined;
-      req.isSmartphoneUser = true;
-
-      await middleware.formRadioButtonChecked(req, res, next);
-
-      expect(res.render).to.have.been.calledWith(
-        `ipv/page/page-ipv-identity-document-start.njk`,
-        sinon.match.has("dcmawEvent", "smartphone"),
-      );
-    });
   });
 
   context("handling pre-dcmaw triage sniffing journey route", () => {
@@ -1078,30 +1052,6 @@ describe("journey middleware", () => {
         csrfToken: sinon.fake(),
         log: { info: sinon.fake(), error: sinon.fake() },
       };
-    });
-
-    it("sets non-smartphone dcmawEvent", async function () {
-      req.method = "GET";
-      req.isSmartphoneUser = false;
-
-      await middleware.handleJourneyPage(req, res, next);
-
-      expect(res.render).to.have.been.calledWith(
-        `ipv/page/pyi-suggest-other-options.njk`,
-        sinon.match.has("dcmawEvent", "dcmaw"),
-      );
-    });
-
-    it("sets smartphone dcmawEvent", async function () {
-      req.method = "GET";
-      req.isSmartphoneUser = true;
-
-      await middleware.handleJourneyPage(req, res, next);
-
-      expect(res.render).to.have.been.calledWith(
-        `ipv/page/pyi-suggest-other-options.njk`,
-        sinon.match.has("dcmawEvent", "smartphone"),
-      );
     });
   });
