+++ src/constants/contexts.js
@@ -1,0 +1,4 @@
+module.exports = Object.freeze({
+  IPHONE: "iphone",
+  ANDROID: "android",
+});
