+++ src/views/ipv/page/page-ipv-identity-document-start.njk
@@ -68,7 +68,7 @@
       },
       items: [
         {
-          value: dcmawEvent,
+          value: "appTriage",
           text: 'pages.pageIpvIdentityDocumentStart.content.formRadioButtons.yes' | translate
         },
         {
