+++ src/views/ipv/page/pyi-cri-escape-no-f2f.njk
@@ -35,7 +35,7 @@
       },
       items: [
         {
-          value: dcmawEvent,
+          value: "appTriage",
           text: 'pages.pyiCriEscapeNoF2f.content.formRadioButtons.continueAppButtonText' | translate,
           hint: {text: 'pages.pyiCriEscapeNoF2f.content.formRadioButtons.continueAppButtonTextHint' | translate}
