+++ src/views/ipv/page/pyi-cri-escape.njk
@@ -36,7 +36,7 @@
       },
       items: [
         {
-          value: dcmawEvent,
+          value: "appTriage",
           text: 'pages.pyiCriEscape.content.formRadioButtons.continueAppButtonText' | translate,
           hint: {text: 'pages.pyiCriEscape.content.formRadioButtons.continueAppButtonTextHint' | translate}
