+++ src/views/ipv/page/pyi-kbv-escape-m2b.njk
@@ -45,7 +45,7 @@
 
         },
         {
-          value: dcmawEvent,
+          value: "appTriage",
           text: 'pages.pyiKbvEscapeM2b.content.formRadioButtons.radioButton2Text' | translate,
           hint: { text: 'pages.pyiKbvEscapeM2b.content.formRadioButtons.radioButton2TextHint' | translate }
