+++ src/views/ipv/page/pyi-suggest-other-options-no-f2f.njk
@@ -35,7 +35,7 @@
       },
       items: [
         {
-          value: dcmawEvent,
+          value: "appTriage",
           text: 'pages.pyiSuggestOtherOptionsNoF2f.content.formRadioButtons.continueAppButtonText' | translate,
           hint: {text: 'pages.pyiSuggestOtherOptionsNoF2f.content.formRadioButtons.continueAppButtonTextHint' | translate}
