+++ src/views/ipv/page/pyi-suggest-other-options.njk
@@ -37,7 +37,7 @@
       },
       items: [
         {
-          value: dcmawEvent,
+          value: "appTriage",
           text: 'pages.pyiSuggestOtherOptions.content.formRadioButtons.continueAppButtonText' | translate,
           hint: {text: 'pages.pyiSuggestOtherOptions.content.formRadioButtons.continueAppButtonTextHint' | translate}
